Allow omitting Content-Type headers
this changes the merge strategy of user provided options so that a
content-type header can be omitted.

i need to not have a content-header, so that <link rel="prefetch" ...
with a GET will work. can't use this with a content-type, or the headers
for my prefetch and request do not match.